backend/requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
billiard==4.2.1
bleach==6.2.0
celery==5.5.2
certifi==2025.4.26
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==45.0.3
daphne==4.2.0
Django==4.2
django-cors-headers==4.7.0
django-cron==0.6.0
django-crontab==0.7.1
django-summernote==0.8.20.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
geographiclib==2.0
geopy==2.4.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
kombu==5.5.3
msgpack==1.1.0
mysqlclient==2.2.7
pillow==11.2.1
prompt_toolkit==3.0.51
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==25.1.0
python-dateutil==2.9.0.post0
pytz==2025.2
redis==6.1.0
requests==2.32.3
service-identity==24.2.0
setuptools==80.8.0
six==1.17.0
sqlparse==0.5.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
zope.interface==7.2
